requirements.txt
pyte==0.8.1
python-dotenv==1.0.0
openai>=1.0.0
requests>=2.28.2
markdown==3.4.4
Jinja2==3.1.2
